Extract individual components and fix file extensions
