+++ src/components/expenseCategoryForm.tsx
@@ -1,0 +1,10 @@
+import { ExpenseTableForm } from '@/components/expenseTableForm';
+import React from 'react';
+
+export function ExpenseCategoryForm() {
+  return (
+    <ExpenseTableForm placeholder="Enter category name..." type="category">
+      New category
+    </ExpenseTableForm>
+  );
+}
