+++ src/components/expenseTable.tsx
@@ -1,215 +1,13 @@
-import React, { PropsWithChildren } from 'react';
-import { ChevronDown, ChevronRight, ChevronUp, Circle, CornerDownRight, Minus, Plus } from 'react-feather';
-import { ExpenseCategory } from '@/models/expenseCategory';
-import { ExpenseItem } from '@/models/expenseItem';
-import { money } from '@/utils/money';
-import { formatDate } from '@/utils/date';
-import { Button, Form, Input } from 'antd';
+import React from 'react';
+import { ExpenseItemForm } from '@/components/expenseItemForm';
+import { ExpenseCategory } from '@/components/expenseCategory';
+import { ExpenseItem } from '@/components/expenseItem';
+import { SortableTableHeader } from '@/components/sortableTableHeader';
+import { ExpenseTotal } from '@/components/expenseTotal';
+import { ExpenseCategoryForm } from '@/components/expenseCategoryForm';
+import { CategoryModel } from '@/models/category.model';
 
-function ExpenseCategory({ name, projected, expenses, children }: PropsWithChildren<ExpenseCategory>) {
-  let [expanded, setExpanded] = React.useState(false);
-  const actual = expenses.reduce((sum, expense) => {
-    return sum + expense.actual;
-  }, 0);
-
-  const font = expanded ? 'font-medium' : 'font-light';
-  const variance = projected - actual;
-
-  const VarianceIcon = variance === 0 ? Minus : Circle;
-  const varianceColor = variance > 0 ? 'text-green-600' : variance === 0 ? 'text-yellow-600' : 'text-red-600';
-  const varianceTitle =
-    variance > 0
-      ? `${money(variance)} bellow projected`
-      : variance === 0
-      ? 'Spot on'
-      : `${money(Math.abs(variance))} above projected`;
-
-  return (
-    <>
-      <tr
-        onClick={() => setExpanded(!expanded)}
-        className="relative cursor-pointer border-b bg-white hover:bg-gray-50 dark:border-gray-700 dark:bg-gray-800 dark:hover:bg-gray-600 [&:last-child]:border-0"
-      >
-        <th
-          scope="row"
-          className={`${font} whitespace-nowrap py-4 pl-6 text-gray-900 dark:text-white [&:last-child]:pr-6`}
-        >
-          <CollapsableTableHeader>{name}</CollapsableTableHeader>
-        </th>
-        <td className="py-4 pl-6 text-right [&:last-child]:pr-6">{money(projected)}</td>
-        <td title={varianceTitle} className="flex flex-row items-center justify-end py-4 pl-6 [&:last-child]:pr-6">
-          {!!actual && <VarianceIcon className={`${varianceColor} mr-auto ml-1 fill-current`} size={10}></VarianceIcon>}
-          {money(actual)}
-        </td>
-      </tr>
-      {expanded && children}
-    </>
-  );
-}
-
-function ExpenseItem({ name, actual, date }: PropsWithChildren<ExpenseItem>) {
-  return (
-    <>
-      <tr className="border-b bg-white hover:bg-gray-50 dark:border-gray-700 dark:bg-gray-800 dark:hover:bg-gray-600 [&:last-child]:border-0">
-        <th
-          colSpan={2}
-          scope="row"
-          className="whitespace-nowrap py-2 pl-6 font-light text-gray-900 dark:text-white [&:last-child]:pr-6"
-        >
-          <div className="flex flex-row items-center gap-2 ">
-            <CornerDownRight className="text-gray-300" size={15} /> {name}
-            <span className="ml-auto text-2xs text-gray-300">{formatDate(date)}</span>
-          </div>
-        </th>
-        <td className="py-2 pl-6 text-right [&:last-child]:pr-6">{money(actual)}</td>
-      </tr>
-    </>
-  );
-}
-
-function CollapsableTableHeader({ children }: PropsWithChildren) {
-  const [collapsed, setCollapsed] = React.useState(true);
-  const Icon = collapsed ? ChevronRight : ChevronDown;
-
-  return (
-    <div onClick={() => setCollapsed(!collapsed)} className="ml-1 flex flex-row items-center">
-      <Icon size={15} className="absolute left-2 text-gray-300" />
-      {children}
-    </div>
-  );
-}
-
-function SortableTableHeader({ children }: PropsWithChildren) {
-  const [direction, setDirection] = React.useState('desc');
-  const Icon = direction === 'desc' ? ChevronDown : ChevronUp;
-
-  return (
-    <div
-      onClick={() => setDirection(direction === 'desc' ? 'asc' : 'desc')}
-      className="flex cursor-pointer flex-row items-center justify-end gap-2"
-    >
-      <Icon size={15} className="text-gray-300" /> {children}
-    </div>
-  );
-}
-
-function ExpenseTotal({ categories }: { categories: ExpenseCategory[] }) {
-  const projected = categories.reduce((sum, category) => {
-    return sum + category.projected;
-  }, 0);
-  const actual = categories.reduce((actual, category) => {
-    return (
-      actual +
-      category.expenses.reduce((sum, expense) => {
-        return sum + expense.actual;
-      }, 0)
-    );
-  }, 0);
-
-  return (
-    <tr className="font-semibold text-gray-900 dark:bg-gray-700 dark:text-white">
-      <th scope="row" className="py-3 pl-6 text-base [&:last-child]:pr-6">
-        Total
-      </th>
-      <td className="py-3 pl-6 text-right [&:last-child]:pr-6">{money(projected)}</td>
-      <td className="py-3 pl-6 text-right [&:last-child]:pr-6">{money(actual)}</td>
-    </tr>
-  );
-}
-
-function ExpenseTableForm({
-  placeholder,
-  type,
-  children,
-}: PropsWithChildren<{ placeholder: string; type: 'category' | 'expense' }>) {
-  const [showForm, setShowForm] = React.useState(false);
-  const formIcon = type === 'category' ? ChevronRight : CornerDownRight;
-
-  const handleEscape = (event: React.KeyboardEvent) => {
-    if (event.key === 'Escape') {
-      setShowForm(false);
-    }
-  };
-  const saveItem = (values: any) => {
-    console.log(values);
-  };
-
-  const FormIcon = showForm ? formIcon : Plus;
-
-  return (
-    <>
-      <tr className="border-b bg-white hover:bg-gray-50 dark:border-gray-700 dark:bg-gray-800 dark:hover:bg-gray-600 [&:last-child]:border-0">
-        <th
-          colSpan={3}
-          scope="row"
-          className="whitespace-nowrap py-2 pl-6 font-light text-gray-900 dark:text-white [&:last-child]:pr-6"
-        >
-          <div className="flex flex-row items-center gap-2">
-            <FormIcon className="text-gray-300" size={15} />
-            {showForm ? (
-              <>
-                <Form onFinish={saveItem} layout="horizontal" className="flex w-full gap-2">
-                  <Form.Item name="name" noStyle>
-                    <Input
-                      autoFocus
-                      bordered={false}
-                      onKeyDown={handleEscape}
-                      placeholder={placeholder}
-                      className="p-0"
-                    ></Input>
-                  </Form.Item>
-                  <Form.Item noStyle>
-                    <Button htmlType="submit" className="py-1 px-2 text-2xs font-light">
-                      Save
-                    </Button>
-                  </Form.Item>
-                  <Button
-                    type="text"
-                    className="py-1 px-2 text-2xs font-light"
-                    onClick={(e) => {
-                      e.preventDefault();
-                      setShowForm(false);
-                    }}
-                  >
-                    Cancel
-                  </Button>
-                </Form>
-              </>
-            ) : (
-              <Button
-                type="link"
-                onClick={(e) => {
-                  e.preventDefault();
-                  setShowForm(!showForm);
-                }}
-                className="flex w-full items-center gap-2 p-0 text-xs font-light text-gray-400"
-              >
-                {children}
-              </Button>
-            )}
-          </div>
-        </th>
-      </tr>
-    </>
-  );
-}
-
-function ExpenseItemForm({ name }: PropsWithChildren<ExpenseCategory>) {
-  return (
-    <ExpenseTableForm placeholder="Expense name..." type="expense">
-      New expense in {name}
-    </ExpenseTableForm>
-  );
-}
-function ExpenseCategoryForm() {
-  return (
-    <ExpenseTableForm placeholder="Enter category name..." type="category">
-      New category
-    </ExpenseTableForm>
-  );
-}
-
-export function ExpenseTable({ categories }: { categories: ExpenseCategory[] }) {
+export function ExpenseTable({ categories }: { categories: CategoryModel[] }) {
   return (
     <div className="relative mx-10 overflow-x-auto shadow-md sm:rounded-lg">
       <table className="w-full text-left text-sm text-gray-500 dark:text-gray-400">
