+++ src/models/category.model.ts
@@ -1,0 +1,8 @@
+import { ItemModel } from '@/models/item.model';
+
+export interface CategoryModel {
+  id: string;
+  name: string;
+  projected: number;
+  expenses: ItemModel[];
+}
