--- src/models/expenseCategory.tsx
@@ -1,8 +1,0 @@
-import { ExpenseItem } from '@/models/expenseItem';
-
-export interface ExpenseCategory {
-  id: string;
-  name: string;
-  projected: number;
-  expenses: ExpenseItem[];
-}
